Fixed Cypress test 116
