+++ cypress/e2e/Building-Overview.cy.ts
@@ -1,27 +1,25 @@
 /**
  * Company user logs in
  * Company user goes to 'My Properties'
- * Company user publishes a new property with a description 
+ * Company user publishes a new property with a description
  * Sign in as public user
  * Public user clicks on new property
  * Public user navigates to building-overview tab
  * Company user logs in
  * Company user deleted newly created property
  */
-
-
-import { CondoType, ParkingType } from 'src/app/models/properties';	
-import { CompanyDTO } from 'src/app/models/users';	
-import { addCondo, addLocker, addParking, login } from './utils.cy';	
-
+import {
+  CreateProperty,
+  DeleteProperty,
+  login,
+} from './utils.cy';
 
 //Company logs in, views 'My Properties' and creates new property
 describe('Company user logs in', () => {
-  beforeEach(() => {	  
+  beforeEach(() => {
     login('rosef12997@hisotyr.com', '123456');
     cy.url().should('eq', 'http://localhost:4200/');
-  });	  
-
+  });
 
   it('Goes to my properties page', () => {
     cy.get('.mat-toolbar.mat-toolbar-single-row').within(() => {
@@ -29,53 +27,17 @@ describe('Company user logs in', () => {
       cy.contains('My Properties').click();
       cy.wait(1000);
     });
-  })
-
-  it('Creates new property',() => {
-    cy.visit('/add-new-property');
-    cy.wait(5000);
-    cy.get('input[name="Name"]').type('Cypress Test 116');
-    cy.get('input[name="Country"]').type('Sample Country');
-    cy.get('input[name="State"]').type('Sample State');
-    cy.get('input[name="City"]').type('Sample City');
-    cy.get('input[formControlName="StreetNN"]').type('Sample Street', {
-      force: true,
-    });
-    cy.get('input[name="ZipCode"]').type('H6J 7K8');
-    cy.get('input[name="Year"]').type('2022');
-    cy.get('textarea[name="Description"]').type('Sample Description');
-    cy.wait(2000);
-    cy.get('input[name="Picture"]').selectFile('./cypress/e2e/TEST.png');
-    cy.get('mat-checkbox[value="Gym"]').click();
-    cy.get('mat-checkbox[value="Pool"]').click();
-    cy.get('mat-checkbox[value="Spa"]').click();
-    cy.get('mat-checkbox[value="Playground"]').click();
-    cy.get('mat-checkbox[value="Meeting Room"]').click();
-    addCondo(CondoType.Sale);
-    addLocker();
-    addParking(ParkingType.Handicap);
-    cy.get('button[type="submit"]').click();
-    cy.get('.loading-indicator').should('exist');
-    cy.wait(5000);
   });
 
-}) 
+  it('Views newly created building', async() => {
+    CreateProperty();
 
-
-//Public user logs in and views newly created property
-describe('Public user logs in', () => {
-  beforeEach(() => {	  
-    login('karinasd07@hotmail.com', 'public_pass');
-    cy.wait(5000);
-    cy.url().should('eq', 'http://localhost:4200/');
-  });	  
-
-  it('Views newly created property', () => {
+    cy.visit('/');
     cy.get('.BuildingDiv').within(() => {
-      cy.contains('.name', 'Cypress Test 116')
-        .parents('.card-content') 
+      cy.contains('.name', 'Sample Building')
+        .parents('.card-content')
         .within(() => {
-          cy.get('#view_btn').click(); 
+          cy.get('#view_btn').click();
         });
     });
     cy.wait(2000);
@@ -84,30 +46,7 @@ describe('Public user logs in', () => {
     cy.get('#general-info').scrollIntoView().should('be.visible');
     cy.get('#facilities').scrollIntoView().should('be.visible');
     cy.get('#company-info').scrollIntoView().should('be.visible');
-  })
 
-}) 
-
-//Company logs in and deletes newly created property
-describe('Company user logs in again', () => {
-  beforeEach(() => {	  
-    login('rosef12997@hisotyr.com', '123456');
-    cy.wait(5000);
-    cy.url().should('eq', 'http://localhost:4200/');
-  });	  
-
-  it('Deletes newly created property', async () => {
-    cy.wait(2000);
-    cy.window().then(async (win) => {
-      const currentUser = (win as any).authService.getUser();
-      const user = (await (win as any).userService.getCompanyUser(
-        currentUser.uid
-      )) as CompanyDTO;
-      console.log('the user', user);
-      const promiseDelete = user.PropertyIds.map(async (ID) => {
-        await(win as any).buildingService.deleteBuilding(ID);
-      });
-      await Promise.all(promiseDelete);
-    });
-  })
-}) 
+    DeleteProperty();
+  });
+});
